backend/requirements.txt
uagents==0.22.5
anthropic==0.54.0
fastapi==0.115.13
uvicorn==0.34.3
python-multipart==0.0.20
pillow==11.2.1
torch==2.7.1
torchvision==0.22.1 
datasets==3.6.0
scikit-learn==1.7.0